+++ packages/react-optimistic/src/index.ts
@@ -1,5 +1,6 @@
 // Re-export all public APIs
 export * from "./useCollection"
+export * from "./useOptimisticMutation"
 export * from "./useLiveQuery"
 export * from "./electric"
